Config to use FQDN or hostname for riemann resmon
By default riemann-resmon was using the FQDN of the server it was
checking as the event hostname. This was slightly annoying as it meant
that went submitting metrics from an application you had to submit as
the FQDN.

This change allows you to select the behaviour, by default it will use
the hostname of the server it is checking.