Add health status to riemann nginx
If the connection to the nginx status page fails, set the health of
nginx to critical. Set it to be ok if the connection to the status page
works.

We've been using this to alert when nginx goes down or the status page
become unavailable. I feel it's a bit nicer this way, rather than
spamming the log with connection refused exceptions :)