Added check and switch for Varnish 4
Fixes #104.

Varnish 4+ changes the field names for Varnish statistics fields.

This change runs varnishstats -V (using popen3 because Varnish outputs
all version information to stderr and not stdout) and captures the
version via a regex.

A case statement switches on the version. A 4 or higher results in using
the new fields. Otherwise it defaults to the Varnish 3.x field names.

The command also had to change because the `-f` option no longer
supports a comma-separated list and now requires a glob or multiple
instantiations.