Improved riemann-proc
Provides the following metrics for running processes:

- process/<pid>-<start-time>/rss
- process/<pid>-<start-time>/vsize
- process/<pid>-<start-time>/running
- process/<pid>-<start-time>/cputime

`description` should contain the command itself.
`state` is `ok` if running or sleeping, `warning` if idle and critical
if zombie or in interruptible wait.
`value` should be either process RSS, VSIZE, CPU TIME or 1 for the /running
case.

This also changes things so that:

- the metric for the number of processes matching will now be called
  `count/<regex>`
- by default the regular expression for matching processes is `".*"` not
  `""`.
- by default the minimum number of processes is set to 0
- by default the maximum number of processes is 32K