Fix cpu and memory usage sorting
When listing processes to fill-in the description of CPU and memory
events, the first line of the output from ps(1) is mixed with the rest
of the output which list processes which can result in weird reports
where the header line is mixed with the rest of the output:

```
root@theremin ~ # ps -eo pcpu,pid,comm | sort -nrb -k1 | head -10
 3.2    2216 corosync
 1.1    2268 pvestatd
%CPU     PID COMMAND
 0.4  460190 fail2ban-server
 0.4    2266 pve-firewall
 0.3    2857 systemd
 0.3    2386 systemd
 0.3    1089 z_rd_int_1
 0.3    1088 z_rd_int_0
 0.2    2138 pmxcfs
```

Add a utility function for sorting lines by reverse numeric value while
keeping the header line untouched.

While here, introduce some CI with GitHub actions.