Make the list of ignored filesystems tunable
There is a bunch of virtual filesystems that do not deserve to be
reported, and they are often operating-system dependent, i.e. FreeBSD
has autofs, devfs, fdescfs, procfs, linprocfs, linsysfs, tmpfs that
always report 100% capacity and can be safely ignored.

This option allows the users to adjust the default ignored filesystems
list if needed (e.g. to un-ignore nfs) and we will be able to use this
list to filter df(1) output on systems without the ability to filter by
filesystem type at the command level.