Add a test to ensure we reports all zfs filesystems
The current code rejects all filesystems which device does not contain a
`/`.  On a ZFS pool, the root filesystem is reported as `zpool` and is
excluded from the reports.

We probably do not want to exclude these filesystems at the root of a
ZFS pool.

Add a unit test to demonstrate the expected behavior.