Add support for FreeBSD system fd monitoring
Rely on the `kern.openfiles` sysctl instead of the presence of lsof(1)
on the system.  Using libffi we could replace the command invocation
with a bunch of calls to libc functions but this would require extra
dependencies which might be premature in this case.  So for now just run
sysctl(1).