Fix Lint/NonLocalExitFromIterator
In some case, it would probably be better to jump to the next iteration
of a loop but I kept the current behaviour of leaving the loop early.

In any case, past ok events from previous iterations would expire when a
metric goes critical and stops processing nodes from a list.