Ignore overlay filesystems by default
In the Linux world, filesystem overlays allows to mount a directory on
top of another directory.  Because we already collect usage information
for the filesystems that contain the "lower" layer, we do not need to
collect information about the "upper" layer which dupplicate the same
information.

This helps avoiding transcient data about short-lived filesystems
created e.g. by CI systems that rely on Docker.