Implement a shared cache for riemann-health
Multiple metrics for MacOS are provided by a single utility and we used
a cache to avoid running it once for each metric.  In order to extend
this logic to other utilities and avoid code duplication, introduce a
"glabal" cache where any function can store data and which is discarded
before each tick.

No functional change.