Add uptime monitoring to riemann-health
Make a new metric with the node uptime available. The metric is the raw
uptime in seconds and the description is a human readable duration.
Critical and Warning thresholds are reversed when compared to other
metrics: a value lower that these thresholds indicates a problem. This
is intended to get notifications when nodes restart unexpectedly.

The uptime check is not enabled by default and must be explicitly
enabled using the --checks flag.