Remove duplicate code
The erlang scripts in the riak_status directory only differ from the
ones in the bin directory by erlang's argument line:
https://www.erlang.org/doc/man/escript.html#export

The ruby script in the riak_status directory seems to be an older
version of the code in the bin directory, not using Riemann Tools and
communicating with some 'vodpod' thing.  Maybe used as a template to
build the riemann-riak tool.

This makes me think the riak_status directory is a leftover of older
files and we can get rid of it.