Normalize class names
Improve the way class and filenames are matched.  The idea is to be able
to automatically require files based on the requested class name using
methods like Rails' String#underscore / String#camelize methods.

This will help with the creation of a generic wrapper abble to load
dynamically extra tools which are unknown to the base riemann-tool gem.

While here, spot and fix a few naming mistakes in previous commit.