Add support for configuration file to riemann-wrapper
If a single argument is passed to riemann-wrapper, assume it is a
configuration file and build ARGV by reading its content.

This is backward compatible with the existing argument parsing and makes
it ealier to run a set of tools with custom configuration for
development or when managing riemann-wrapper config with configuration
management.