Improve zpool state reporting
When a FAULTED or DEGRADED zpool is back ONLINE and resilvering after
changing disks, it is still reported in `zpool status -x` output until
the pool is healthy again.  Such a state should probably not be reported
as critical but it is still something to keep an eye on, so report all
the known-bad states as `critial` as before but report this resilvering
status (and also unknown statuses that might be introduced in the
future) as `warning` only.