Fix parsing of mdstat when device is being checked
The content of /dev/mdstat when a device is being checked looks like the
one of a device being recovering, and [a few more values can be
reported](https://github.com/torvalds/linux/blob/7726d4c3e60bfe206738894267414a5f10510f1a/drivers/md/md.c#L8089-L8098).

This information is reported *before* the bitmap when present, and not
after as the parser expected (the examples from the documentation did
not mix bitmaps and recovery, but the weekly check from Debian broke the
`mdstat` reporting during the check).

Fix the ordering and allow the new values to be parsed.  Add an example
from a live output of one of our nodes to ensure we don't break this in
the future.