Improve error reporting on parse error
When a parser cannot parse a document, it raised a Racc::ParseError if
the document was malformed or a RuntimeError if the content was not
processable.  In any case, the exception was not explicitely catched and
was logged to stderr.  If events where previously reported they would
eventualy expire, otherwise nothing would make the failure visible on the
riemann side.

Improve exception handling by formatting a custom error messages which
show the line of the failure in the parsed document and sending it as a
critical event to riemann.