Add `riemann-http-check` to monitor HTTP(S) resources
riemann-tools feature `riemann-portcheck` which can check if a port can
be reached, but this is intended to check local services only and does
not communicate over the connected socket, so can't detect some failure
conditions like bad gateways.

Add a riemann-http-check that accept a list of URI of resources to check
using HTTP(S) GET requests. For each URI, resolve the IP addresses that
provide the service, and for each IP address generate 3 events:
connection latency, response latency and response code. Then generate a
last consistency event.

 1. connection latency is a duration in seconds, the warning and
    critical threshold are respectively 100ms and 250ms, they can be
    tuned with the --connection-latency-warning and
    --connection-latency-critical flags;
 2. response latency is a duration in seconds, the warning and critical
    threshold are respectively 500ms and 1000ms, they can be tuned with
    the --response-latency-warning and --response-latency-critical
    flags;
 3. response code is the numeric response code and is looked for in a
    list of acceptable results, by default 200 OK, 301 Moved Permanently
    and 401 Unauthorized (other codes are considered to be non-working
    remote services);
 4. consistency reports an 'ok' state if all responses have the same
    response code and body.

When using HTTPS URI, the validity of certificates is not checked and
any TLS error is silently ignored. The idea is to only check for
resource availability with this tool and have another tool to check for
TLS connections in a generic — protocol agnostic — way.