Add resolver and worker threads to http-check
IP addresses resolution and HTTP resources fetching are I/O bound. When
testing a large amount of HTTP endpoints, this can be an issue.  Some
measurement we did for a use case show that checking 112 URI (most of
them being serviced by 2 IP addresses) took 95 seconds.

Implement an address resolution queue and a working queue in order to
parallelize tasks in multiple threads.  With a default of 5 resolvers
and 20 workers, our test case now completes in about 4 seconds, which is
more than 23 times faster.  These values can be tuned using the
--resolvers and --workers new parameters.

While here, add metrics to track the resolvers and workers utilization and
saturation.