Fix `riemann-http-check` with unresolvable domains
When the host part of an URI cannot be resolved, the
`riemann-http-check' assume it is an IP address and attempt to parse it.
But when this host is not an IP address, a IPAddr::InvalidAddressError
exception is raised and is not catch, no metric is emitted, and the
thread that raised the exception terminate.  On next attempt, another
thread is killed, and soon no more resolver thread is available for the
system to function properly.

Catch the IPAddr::InvalidAddressError exception and ignore it.  When the
name is not resolvable, this produce a consistency metric with the
critical state because no response can be collected for the URI.