Send events in bulk when they are stacking
Sending an event to riemann can take some time. If during this time
multiple events are generated, they will stack and be send one by one
instead of being transmitted in bulk.

We added a new API function to riemann-client to send events in bulk:
https://github.com/riemann/riemann-ruby-client/pull/44

Add a send queue and a worker thread that will send events in bulk if
more than one is ready to be sent.