Add redirect support to `riemann-http-check`
When we get a redirect response, and the target has the same origin than
the initial URI (same scheme, host and port), initiate a new request to
that target route and report it like we do for the initial page.

Enable this feature by default, but don't allow more that 5 redirects to
prevent infinite querying when monitoring broken redirected pages (a
critial event is however returned in this case).