Use truncated exponential backoff for reconnection
When the communication with the Riemann server is connected (TCP/TLS)
and the link breaks, riemann-wrapper drop the events that failed to be
send, log a warning, and immediately try to reconnect and proceed with
remaining data.

When the Riemann server is unreachable because of some network
connectivity issue, this new connection will likely immediately fail,
freshly gathered events will be dropped, a new warning will be logged,
and a new connection will be tried immediately.

Because we do not wait before reconnecting, we log an unexpectedly large
amount of information about dropped messages, and because of the delays
introduced by the reconnection attempts, we might be sending stale data
when the connection succeed again.

Rework the disconnection detection logic to apply some truncated
exponential backoff when the connection dies.  Sleep at least 0.5 and at
most 30s between attempts, and drop any pending events before trying to
reconnect.