Add support for options with spaces to `riemann-wrapper`
When `riemann-wrapper` is used with a configuration file (and only in
this case), it is not possible to prevent splitting of options
containing spaces, adding an arbitrary constraint for the user if they
want to pass options with spaces (e.g. when passing a custom format
string to a tool).

Rework options splitting when reading options from a file, so that the
user can:
* quote part of the string (i.e. `options: --format='foo bar'`);
* escape spaces (i.e. `options: --format=foo\ bar`).

The above examples are evaluated as an array with a single string item:
`["--format=foo bar"]`.