Allow using the latest version of rack (3.x)
The project dependencies now allow rack 3 to be installed.  This
backwards incompatible version extracted handlers in a new `rackup` gem.
While we could indicate we prefer to use rack 2 which is still receiving
update, it is more future-proof to update our code to be compatible with
the latest version of the dependency.

This is a development dependency only so will not affect users of the
project.