Terminate `Riemann::Tools::HttpCheck` threads between tests
Under normal operation, we have a single instance of the
`Riemann::Tools::HttpCheck` class for the lifetime of the monitoring
process.

But when testing, we create a new instance for each test, each with its
resolvers and worker thread pools.  The test process will have more and
more threads, until it eventually hit the OS limit of the maximum number
of threads for a process and cause an exception:

> ThreadError: can't create Thread: Resource temporarily unavailable

Make sure that the threads are terminated at the end of each test to
avoid running out of resources if they are limited.