Fix Linux memory reporting when using ZFS
On Linux, memory used by the ZFS ARC is reported as used while in fact
it mostly consist of cached data that can be reclaimed by the operating
system if necessary.

If the system use ZFS, gather the ARC statistics to compute the size
of the ARC that can be evicted and do not count this as used memory.
This fix the obviously wrong reporting we can see on systems running
ZFS.

Containers running on nodes with ZFS can access the hosts ARC
statistics from the container, resulting in wrong memory reporting.  To
avoid this, we detect if we are running from a container and skip
reading the ARC statistics in that case.