Fix mixing `#opts` and `#options`
`#opts` is an alias for `#options` in the `Riemann::Tools` class.

While most of the tools use `#opts`, a few ones une `#options`, but the
files in this commit use both functions, which can cause confusion.

To avoid this, stick to `#opts`.  Files that consitently use `#options`
are kept as is.

No functional change.