Detect stray arguments
When a tool is run with stray arguments, they are silently ignored.
This typically happen when a parameter must be repeated for each value
(e.g. the `--tag` parameter is passed multiple values with `--tag foo
--tag bar` and passing `--tag foo bar` only provide a single tag and the
`bar` value is an argument that is silently ignored).

In order to avoid this issue, raise an error by default if stray
arguments are found.  It is possible to indicate that stray arguments
are allowed by expilcitly saying so.  In this case, the arguments are
accessible using the `#argv` method.