Add support for a minimum TTL for events
When running with riemann-wrappers, one may want to run different tools
with distinct ttl, but the `--ttl` flag cannot be specified twice, so
the following configuration will not work:

```
---
options: --ttl 300
tools:
- name: "health"
- name: "rdap"
  options: "--interval 21600 --ttl 86400 --domains example.com example.net"
```

A workaround is to add an explicit ttl for each tool, but it is not
convenient.

Introduce a `--minimum-ttl` flag that can be used as a global option to
circumvent this issue.