Add support for ignoring IPs by ASN in `riemann-http`
Monitoring a website behind a Content Delivery Network (CDN) may lead to
flapping metrics when the short-lived IP addresses where the service is
accessible change.

Allow to provide a list of Autonomous System Numbers (ASN) that we can
ignore for well-known CDN service providers.  Use the MaxMind ASN
database provided by the user for IP lookups.

This is not a hard dependency as no ASN filtering is done by default, so
only add this dependency for testing and assume the end-user will handle
the soft requirement on his own if they want to do filter-out some ASN.