Improve `riemann-hwmon` reliability
Some old hardware do not report CPU package temperature and the creation
of `temp1_*` (reserved for package temperature) files for the CPU is
skipped.  When this happen, other `temp2_*`, `temp3_*` etc. files are
available as usual for each core temperature.

Because we tried to enumerate the possible device names, we failed to
catpure these sensor values.

Moreover, people report on the internet "holes" in the numbering of the
hwmon files, skipping from 8 to 10.

In order to avoid these issues, rely on globbing to find all appropriate
files rather than on enumerations.