Allow test webserver to be reached with any hostname
Recent versions of Sinatra ensure the requested hostname is allowed and
return a 403 error if not.  This trigger in CI because we use custom
host names rather than localhost.

See:
https://github.com/sinatra/sinatra/pull/2053

Allow any domain as this is used for testing and we are not concerned by
the security vulnerability this change upstream is meant to fix.