Distinguish i2c devices in riemann-hwmon
When multiple identical i2c devices are present on a host (typical case
for temperature sensors on memory modules), they overwrite each other
metrics making monitoring unreliable.

Detect if the hwmon sensor is queried over i2c and if so add location
information so that each device is reported with a different service.