Ignore null-critical value for `riemann-hwmon`
Some devices wrongly report 0 as a critical value (e.g. RAM temperature
sensors).  The critical state of such events is wrong, and we cannot
count on it.  As only a temperature can possibly make sense when passing
above 0 (all other metrics are unsigned ones), and such condition does
not really make sense in a computing environment, this should not have
bad consequences for end users.